refactor: Moved forms to new folders
