+++ src/Form/Areas/StateType.php
@@ -1,28 +1,25 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Areas;
 
-use App\DataTransferObjects\LoginTypeDTO;
+use App\Entity\State;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\PasswordType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class LoginType extends AbstractType
+class StateType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('username', TextType::class)
-            ->add('password', PasswordType::class)
+            ->add('name')
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => LoginTypeDTO::class,
+            'data_class' => State::class,
         ]);
     }
 }
