+++ src/Form/Security/LoginType.php
@@ -1,25 +1,28 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Security;
 
-use App\Entity\State;
+use App\DataTransferObjects\LoginTypeDTO;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class StateType extends AbstractType
+class LoginType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('name')
+            ->add('username', TextType::class)
+            ->add('password', PasswordType::class)
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => State::class,
+            'data_class' => LoginTypeDTO::class,
         ]);
     }
 }
