+++ src/Form/Security/RegistrationType.php
@@ -1,36 +1,45 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Security;
 
-use App\Validator\UserPasswordConstraint;
+use App\DataTransferObjects\RegisterTypeDTO;
+use App\Form\DataMapper\RegistrationDataMapper;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotBlank;
 
-class UpdateUserPasswordType extends AbstractType
+class RegistrationType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('currentPassword', PasswordType::class, [
+            ->add('username', TextType::class)
+            ->add('email', EmailType::class)
+            ->add('agreeTerms', CheckboxType::class, [
                 'mapped' => false,
                 'constraints' => [
-                    new UserPasswordConstraint(),
+                    new IsTrue([
+                        'message' => 'You have to agree to our terms.',
+                    ]),
                 ],
             ])
             ->add('password', RepeatedType::class, [
                 'type' => PasswordType::class,
-                'label' => 'New Password',
+                'mapped' => false,
                 'attr' => ['autocomplete' => 'new-password'],
                 'invalid_message' => 'The password fields must match.',
                 'options' => ['attr' => ['class' => 'password-field']],
                 'required' => true,
-                'first_options' => ['label' => 'New Password'],
-                'second_options' => ['label' => 'Repeat new Password'],
+                'first_options' => ['label' => 'Password'],
+                'second_options' => ['label' => 'Repeat Password'],
                 'constraints' => [
                     new NotBlank([
                         'message' => 'Please enter a password',
@@ -44,12 +53,14 @@ class UpdateUserPasswordType extends AbstractType
                 ],
             ])
         ;
+
+        $builder->setDataMapper(new RegistrationDataMapper());
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            // Configure your form options here
+            'data_class' => RegisterTypeDTO::class,
         ]);
     }
 }
