+++ src/Form/User/UpdateUserPasswordType.php
@@ -1,45 +1,36 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\User;
 
-use App\DataTransferObjects\RegisterTypeDTO;
-use App\Form\DataMapper\RegistrationDataMapper;
+use App\Validator\UserPasswordConstraint;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotBlank;
 
-class RegistrationType extends AbstractType
+class UpdateUserPasswordType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('username', TextType::class)
-            ->add('email', EmailType::class)
-            ->add('agreeTerms', CheckboxType::class, [
+            ->add('currentPassword', PasswordType::class, [
                 'mapped' => false,
                 'constraints' => [
-                    new IsTrue([
-                        'message' => 'You have to agree to our terms.',
-                    ]),
+                    new UserPasswordConstraint(),
                 ],
             ])
             ->add('password', RepeatedType::class, [
                 'type' => PasswordType::class,
-                'mapped' => false,
+                'label' => 'New Password',
                 'attr' => ['autocomplete' => 'new-password'],
                 'invalid_message' => 'The password fields must match.',
                 'options' => ['attr' => ['class' => 'password-field']],
                 'required' => true,
-                'first_options' => ['label' => 'Password'],
-                'second_options' => ['label' => 'Repeat Password'],
+                'first_options' => ['label' => 'New Password'],
+                'second_options' => ['label' => 'Repeat new Password'],
                 'constraints' => [
                     new NotBlank([
                         'message' => 'Please enter a password',
@@ -53,14 +44,12 @@ class RegistrationType extends AbstractType
                 ],
             ])
         ;
-
-        $builder->setDataMapper(new RegistrationDataMapper());
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => RegisterTypeDTO::class,
+            // Configure your form options here
         ]);
     }
 }
